requirements.txt
Pillow>=9.0
pystray>=0.19
pywin32>=305; sys_platform == 'win32'
